add: details to card and minor refactor
